+++ math_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestPow(t *testing.T) {
 
-	assert.Equal(t, 16, Pow(2, 4))
-	assert.Equal(t, 65536, Pow(2, 16))
-	assert.Equal(t, 1048576, Pow(2, 20))
+	assert.Equal(t, int64(16), Pow(2, 4))
+	assert.Equal(t, int64(65536), Pow(2, 16))
+	assert.Equal(t, int64(1048576), Pow(2, 20))
 }
